requirements.txt
Flask==3.1.0
numpy==2.2.0
paddleocr
pandas==2.2.3
Pillow==11.0.0
protobuf==5.29.2
python-dotenv==1.0.1
python_magic==0.4.27
google-ai-generativelanguage==0.6.10
google-api-core==2.23.0
google-api-python-client==2.153.0
google-auth==2.36.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.3
googleapis-common-protos==1.66.0
google-genai==1.3.0
